% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_console_redirect.R
\name{withConsoleRedirect}
\alias{withConsoleRedirect}
\title{withConsoleRedirect}
\usage{
withConsoleRedirect(containerId, expr)
}
\arguments{
\item{containerId}{the id of the container}

\item{expr}{an R expression}
}
\description{
withConsoleRedirect
}
